Change in post section and login section

+++ controllers/post.controller.js
@@ -1,17 +1,17 @@
 import Post from "../models/post.model.js";
-import cloudinary from "../utils/Cloudinary.js"
-import fs from 'fs';
+import cloudinary from "../utils/Cloudinary.js";
+import fs from "fs";
 
 const createPost = async (req, res) => {
   try {
-    const file = req.file
+    const file = req.file;
 
     if (!file) {
-      return res.status(400).json({ error: 'No image uploaded' });
+      return res.status(400).json({ error: "No image uploaded" });
     }
 
     const result = await cloudinary.uploader.upload(file.path, {
-      folder: 'posts'
+      folder: "posts",
     });
 
     // Remove local file after uploading
@@ -20,16 +20,18 @@ const createPost = async (req, res) => {
     // Save post
     const newPost = new Post({
       title: req.body.title,
+      description: req.body.description,
       imageUrl: result.secure_url,
-      imageId: result.public_id, 
+      imageId: result.public_id,
+      userId: req.user._id,
     });
 
-    await newPost.save()
+    await newPost.save();
 
-    res.status(201).json(newPost)
+    res.status(201).json(newPost);
   } catch (err) {
-    console.error(err)
-    res.status(500).json({ error: 'Internal server error' })
+    console.error(err);
+    res.status(500).json({ error: "Internal server error" });
   }
 };
 
@@ -42,12 +44,16 @@ const deletePost = async (req, res) => {
       return res.status(404).json({ error: "Post not found" });
     }
 
-    // Delete image from Cloudinary
+    if (post.userId.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ error: "Unauthorized to delete this post" });
+    }
+
     if (post.imageId) {
       await cloudinary.uploader.destroy(post.imageId);
     }
 
-    // Delete post from DB
     await Post.findByIdAndDelete(id);
 
     res.status(200).json({ message: "Post deleted successfully" });
@@ -57,7 +63,22 @@ const deletePost = async (req, res) => {
   }
 };
 
-export {
-  createPost,
-  deletePost
-}
+const getMyPosts = async (req, res) => {
+  try {
+    const posts = await Post.find({ userId: req.user._id });
+    res.status(200).json(posts);
+  } catch (err) {
+    res.status(500).json({ error: "Something went wrong" });
+  }
+};
+
+const getAllPosts = async (req, res) => {
+  try {
+    const posts = await Post.find().populate("userId", "username email");
+    res.status(200).json(posts);
+  } catch (err) {
+    res.status(500).json({ error: "Something went wrong" });
+  }
+};
+
+export { createPost, deletePost, getAllPosts, getMyPosts };
